Revert "Merge branch 'Skill' into dev"
This reverts commit 072aaeb74be5b52b48819586fd6a5e97cfd20ef6, reversing
changes made to 79308bd8995de631145476585300fa191e0c5ba9.

+++ Assets/01.Scripts/Interface/ICharacter.cs
@@ -6,9 +6,9 @@ public interface ICharacter
 {
     string Name { get; set; }
     int Level { get; set; }
-    float AttackDamage { get; set; }
+    int AttackDamage { get; set; }
     float MoveSpeed { get; set; }
-    float CurrentHealth { get; set; }
-    float MaxHealth { get; set; }
+    int CurrentHealth { get; set; }
+    int MaxHealth { get; set; }
     float Experience { get; set; }
 }
